requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
